# src/l1/piccolod/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Piccolo OS Admin API (Draft)
  version: 0.1.0
security:
  - cookieAuth: []
servers:
  - url: /api/v1
paths:
  /apps:
    get:
      summary: List installed apps
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseApps'
    post:
      summary: Install or update an app from app.yaml
      requestBody:
        required: true
        content:
          application/x-yaml:
            schema:
              type: string
          application/json:
            schema:
              type: object
              properties:
                app_definition:
                  type: string
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseApp'
        '400': { description: Bad Request, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /apps/validate:
    post:
      summary: Validate an app.yaml without installing
      requestBody:
        required: true
        content:
          application/x-yaml:
            schema: { type: string }
          application/json:
            schema:
              type: object
              properties:
                app_definition:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid: { type: boolean }
        '400': { description: Bad Request, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /apps/{name}:
    get:
      summary: Get app details (with services)
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseAppWithServices'
    delete:
      summary: Uninstall app
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
        - in: query
          name: purge
          schema: { type: boolean }
      responses:
        '200': { description: OK }
  /apps/{name}/start:
    post:
      summary: Start app
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /apps/{name}/stop:
    post:
      summary: Stop app
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /apps/{name}/update:
    post:
      summary: Update an app to a newer tag
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                tag: { type: string, nullable: true }
      responses:
        '200': { description: OK }
        '409': { description: Conflict, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /apps/{name}/revert:
    post:
      summary: Revert an app to previous version
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /apps/{name}/logs:
    get:
      summary: Get recent logs for an app
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
        - in: query
          name: since
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AppLogs' }
        '404': { description: Not Found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
...
  /services/{name}:
    get:
      summary: Get a single service endpoint by name
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  service: { $ref: '#/components/schemas/ServiceEndpoint' }
...
  /crypto/setup:
    post:
      summary: Initialize crypto using admin password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password: { type: string }
      responses:
        '200': { description: OK }
        '400': { description: Bad Request, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /crypto/unlock:
    post:
      summary: Unlock crypto with admin password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password: { type: string }
                recovery_key: { type: string, description: Space-separated 24-word mnemonic }
      responses:
        '200': { description: OK }
        '401': { description: Unauthorized }
...
  /crypto/recovery-key/generate:
    post:
      summary: Generate recovery key (requires unlocked)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                password: { type: string, description: Optional password to unlock transiently }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  words:
                    type: array
                    items: { type: string }
...
  /remote/configure:
    post:
      summary: Configure Nexus remote access
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                endpoint: { type: string }
                device_id: { type: string }
                device_secret: { type: string }
                hostname: { type: string }
                acme_directory: { type: string, description: ACME directory URL for staging/tests }
      responses:
        '200': { description: OK }
        '400': { description: Bad Request, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '429': { description: Too Many Requests }
...
  /catalog/{name}/template:
    get:
      summary: Get example app.yaml for a catalog item
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/x-yaml:
              schema:
                type: string
        '404': { description: Not Found }
  /auth/setup:
    post:
      summary: First-run admin setup
      responses:
        '200': { description: OK }
        '400': { description: Bad Request }
  /auth/login:
    post:
      summary: Login
      responses:
        '200': { description: OK }
        '401': { description: Unauthorized }
        '429': { description: Too Many Requests, headers: { Retry-After: { schema: { type: integer } } } }
...
  /auth/password:
    post:
      summary: Change admin password (rewrap keys)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                old_password: { type: string }
                new_password: { type: string }
      responses:
        '200': { description: OK }
        '400': { description: Bad Request }
        '401': { description: Unauthorized }
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        code: { type: integer }
        message: { type: string }
        details: { type: string, nullable: true }
        next_step: { type: string, nullable: true }
    ResponseApps:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/App' }
        message: { type: string }
    ResponseApp:
      type: object
      properties:
        data: { $ref: '#/components/schemas/App' }
        message: { type: string }
    ResponseAppWithServices:
      type: object
      properties:
        data:
          type: object
          properties:
            app: { $ref: '#/components/schemas/App' }
            services:
              type: array
              items: { $ref: '#/components/schemas/ServiceEndpoint' }
    App:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        image: { type: string }
        subdomain: { type: string }
        type: { type: string }
        status: { type: string, enum: [running, stopped, error] }
        volumes: { type: array, items: { $ref: '#/components/schemas/AppVolume' } }
        environment: { type: object, additionalProperties: { type: string } }
    AppVolume:
      type: object
      properties:
        container: { type: string }
        host: { type: string }
        size_limit: { type: string }
    ServiceEndpoint:
      type: object
      properties:
        app: { type: string }
        subdomain: { type: string }
        name: { type: string }
        guest_port: { type: integer }
        host_port: { type: integer, description: Port bound on 127.0.0.1 for LAN }
        public_port: { type: integer }
        flow: { type: string }
        protocol: { type: string }
        middleware: { type: array, items: { type: object } }
        local_url: { type: string, nullable: true }
...
    AppLogs:
      type: object
      properties:
        app: { type: string }
        entries:
          type: array
          items:
            type: object
            properties:
              ts: { type: string, format: date-time }
              level: { type: string }
              message: { type: string }
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: piccolo_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token